requirements.txt
opencv-python==4.6.0.66
face-recognition==1.3.0
streamlit==1.14.1
numpy==1.23.4
Pillow==9.2.0